+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed an issue in **update-release-notes** where suggestion fix failed in validation.
 * Fixed a bug where `.env` files didn't load. @nicolas-rdgs
 * Fixed a bug where **validate** command failed when the *categories* field in the pack metadata was empty for non-integration packs.
+* Added *system* and *item-type* arguments to **download** command to support downloading of system items.
 
 # 1.6.1
 * Added the '--use-packs-known-words' argument to the **doc-review** command